[Decomp][ObjectMgr] Cohesion split: extract text loaders to ObjectMgrText.cpp
