+++ src/game/Object/ObjectMgr.cpp
@@ -4571,56 +4571,6 @@ void ObjectMgr::LoadPetCreateSpells()
     sLog.outString(">> Loaded %u pet create spells from table and %u from DBC", count, dcount);
 }
 
-/**
- * @brief Loads page text records and validates page chains.
- */
-void ObjectMgr::LoadPageTexts()
-{
-    sPageTextStore.Load();
-    sLog.outString(">> Loaded %u page texts", sPageTextStore.GetRecordCount());
-    sLog.outString();
-
-    for (uint32 i = 1; i < sPageTextStore.GetMaxEntry(); ++i)
-    {
-        // check data correctness
-        PageText const* page = sPageTextStore.LookupEntry<PageText>(i);
-        if (!page)
-        {
-            continue;
-        }
-
-        if (page->Next_Page && !sPageTextStore.LookupEntry<PageText>(page->Next_Page))
-        {
-            sLog.outErrorDb("Page text (Id: %u) has not existing next page (Id:%u)", i, page->Next_Page);
-            continue;
-        }
-
-        // detect circular reference
-        std::set<uint32> checkedPages;
-        for (PageText const* pageItr = page; pageItr; pageItr = sPageTextStore.LookupEntry<PageText>(pageItr->Next_Page))
-        {
-            if (!pageItr->Next_Page)
-            {
-                break;
-            }
-            checkedPages.insert(pageItr->Page_ID);
-            if (checkedPages.find(pageItr->Next_Page) != checkedPages.end())
-            {
-                std::ostringstream ss;
-                ss << "The text page(s) ";
-                for (std::set<uint32>::iterator itr = checkedPages.begin(); itr != checkedPages.end(); ++itr)
-                {
-                    ss << *itr << " ";
-                }
-                ss << "create(s) a circular reference, which can cause the server to freeze. Changing Next_Page of page "
-                   << pageItr->Page_ID << " to 0";
-                sLog.outErrorDb("%s", ss.str().c_str());
-                const_cast<PageText*>(pageItr)->Next_Page = 0;
-                break;
-            }
-        }
-    }
-}
 
 
 struct SQLInstanceLoader : public SQLStorageLoaderBase<SQLInstanceLoader, SQLStorage>
@@ -4761,65 +4711,6 @@ GossipText const* ObjectMgr::GetGossipText(uint32 Text_ID) const
     return NULL;
 }
 
-/**
- * @brief Loads npc gossip text records from the database.
- */
-void ObjectMgr::LoadGossipText()
-{
-    QueryResult* result = WorldDatabase.Query("SELECT * FROM `npc_text`");
-
-    int count = 0;
-    if (!result)
-    {
-        BarGoLink bar(1);
-        bar.step();
-
-        sLog.outString(">> Loaded %u npc texts", count);
-        sLog.outString();
-        return;
-    }
-
-    BarGoLink bar(result->GetRowCount());
-
-    do
-    {
-        ++count;
-        int cic = 0;
-
-        Field* fields = result->Fetch();
-
-        bar.step();
-
-        uint32 Text_ID    = fields[cic++].GetUInt32();
-        if (!Text_ID)
-        {
-            sLog.outErrorDb("Table `npc_text` has record wit reserved id 0, ignore.");
-            continue;
-        }
-
-        GossipText& gText = mGossipText[Text_ID];
-
-        for (int i = 0; i < MAX_GOSSIP_TEXT_OPTIONS; ++i)
-        {
-            gText.Options[i].Text_0           = fields[cic++].GetCppString();
-            gText.Options[i].Text_1           = fields[cic++].GetCppString();
-
-            gText.Options[i].Language         = fields[cic++].GetUInt32();
-            gText.Options[i].Probability      = fields[cic++].GetFloat();
-
-            for (int j = 0; j < 3; ++j)
-            {
-                gText.Options[i].Emotes[j]._Delay  = fields[cic++].GetUInt32();
-                gText.Options[i].Emotes[j]._Emote  = fields[cic++].GetUInt32();
-            }
-        }
-    }
-    while (result->NextRow());
-
-    sLog.outString(">> Loaded %u npc texts", count);
-    sLog.outString();
-    delete result;
-}
 
 
 // not very fast function but it is called only once a day, or on starting-up
